+++ tests/integration/components/power-select/public-actions-test.js
@@ -6,6 +6,7 @@ import { numbers } from '../constants';
 
 function assertPublicAPIShape(assert, select) {
   assert.equal(typeof select.isOpen, 'boolean', 'select.isOpen is a boolean');
+  assert.equal(typeof select.disabled, 'boolean', 'select.disabled is a boolean');
   assert.equal(typeof select.isActive, 'boolean', 'select.isActive is a boolean');
   assert.equal(typeof select.loading, 'boolean', 'select.loading is a boolean');
   assert.ok(select.options instanceof Array, 'select.options is an array');
@@ -31,7 +32,7 @@ moduleForComponent('ember-power-select', 'Integration | Component | Ember Power 
 });
 
 test('The search action of single selects action receives the search term and the public API', function(assert) {
-  assert.expect(20);
+  assert.expect(21);
 
   this.numbers = numbers;
   this.handleSearch = (term, select) => {
@@ -50,7 +51,7 @@ test('The search action of single selects action receives the search term and th
 });
 
 test('The search action of multiple selects action receives the search term and the public API', function(assert) {
-  assert.expect(20);
+  assert.expect(21);
 
   this.numbers = numbers;
   this.handleSearch = (term, select) => {
@@ -69,7 +70,7 @@ test('The search action of multiple selects action receives the search term and 
 });
 
 test('The onchange of single selects action receives the selection and the public API', function(assert) {
-  assert.expect(20);
+  assert.expect(21);
 
   this.numbers = numbers;
   this.handleChange = (selected, select) => {
@@ -88,7 +89,7 @@ test('The onchange of single selects action receives the selection and the publi
 });
 
 test('The onchange of multiple selects action receives the selection and the public API', function(assert) {
-  assert.expect(20);
+  assert.expect(21);
 
   this.numbers = numbers;
   this.handleChange = (selected, select) => {
@@ -107,7 +108,7 @@ test('The onchange of multiple selects action receives the selection and the pub
 });
 
 test('The onkeydown of single selects action receives the public API and the keydown event when fired on the searchbox', function(assert) {
-  assert.expect(20);
+  assert.expect(21);
 
   this.numbers = numbers;
   this.onKeyDown = (select, e) => {
@@ -151,7 +152,7 @@ test('The onkeydown can be used to easily allow to select on tab', function(asse
 });
 
 test('The onkeydown of multiple selects action receives the public API and the keydown event', function(assert) {
-  assert.expect(20);
+  assert.expect(21);
 
   this.numbers = numbers;
   this.onKeyDown = (select, e) => {
@@ -170,7 +171,7 @@ test('The onkeydown of multiple selects action receives the public API and the k
 });
 
 test('returning false from the `onkeydown` action prevents the default behaviour in single selects', function(assert) {
-  assert.expect(44);
+  assert.expect(46);
 
   this.numbers = numbers;
   this.handleKeyDown = (select, e) => {
@@ -193,7 +194,7 @@ test('returning false from the `onkeydown` action prevents the default behaviour
 });
 
 test('returning false from the `onkeydown` action prevents the default behaviour in multiple selects', function(assert) {
-  assert.expect(22);
+  assert.expect(23);
 
   this.numbers = numbers;
   this.handleKeyDown = (select, e) => {
@@ -214,7 +215,7 @@ test('returning false from the `onkeydown` action prevents the default behaviour
 });
 
 test('The onfocus of single selects action receives the public API and the focus event', function(assert) {
-  assert.expect(20);
+  assert.expect(21);
 
   this.numbers = numbers;
   this.handleFocus = (select, e) => {
@@ -232,7 +233,7 @@ test('The onfocus of single selects action receives the public API and the focus
 });
 
 test('The onfocus of multiple selects action receives the public API and the focus event', function(assert) {
-  assert.expect(20);
+  assert.expect(21);
 
   this.numbers = numbers;
   this.handleFocus = (select, e) => {
@@ -250,7 +251,7 @@ test('The onfocus of multiple selects action receives the public API and the foc
 });
 
 test('the `onopen` action is invoked just before the dropdown opens', function(assert) {
-  assert.expect(22);
+  assert.expect(23);
 
   this.numbers = numbers;
   this.handleOpen = (select, e) => {
@@ -270,7 +271,7 @@ test('the `onopen` action is invoked just before the dropdown opens', function(a
 });
 
 test('returning false from the `onopen` action prevents the single select from opening', function(assert) {
-  assert.expect(22);
+  assert.expect(23);
 
   this.numbers = numbers;
   this.handleOpen = (select, e) => {
@@ -291,7 +292,7 @@ test('returning false from the `onopen` action prevents the single select from o
 });
 
 test('returning false from the `onopen` action prevents the multiple select from opening', function(assert) {
-  assert.expect(22);
+  assert.expect(23);
 
   this.numbers = numbers;
   this.handleOpen = (select, e) => {
@@ -312,7 +313,7 @@ test('returning false from the `onopen` action prevents the multiple select from
 });
 
 test('the `onclose` action is invoked just before the dropdown closes', function(assert) {
-  assert.expect(22);
+  assert.expect(23);
 
   this.numbers = numbers;
   this.handleClose = (select, e) => {
@@ -333,7 +334,7 @@ test('the `onclose` action is invoked just before the dropdown closes', function
 });
 
 test('returning false from the `onclose` action prevents the single select from closing', function(assert) {
-  assert.expect(23);
+  assert.expect(24);
 
   this.numbers = numbers;
   this.handleClose = (select, e) => {
@@ -356,7 +357,7 @@ test('returning false from the `onclose` action prevents the single select from 
 });
 
 test('returning false from the `onclose` action prevents the multiple select from closing', function(assert) {
-  assert.expect(23);
+  assert.expect(24);
 
   this.numbers = numbers;
   this.handleClose = (select, e) => {
@@ -379,7 +380,7 @@ test('returning false from the `onclose` action prevents the multiple select fro
 });
 
 test('the `oninput` action is invoked when the user modifies the text of the search input on single selects, and the search happens', function(assert) {
-  assert.expect(26);
+  assert.expect(27);
 
   this.numbers = numbers;
   this.handleInput = (value, select, e) => {
@@ -404,7 +405,7 @@ test('the `oninput` action is invoked when the user modifies the text of the sea
 });
 
 test('the `oninput` action is invoked when the user modifies the text of the search input on multiple selects, and the search happens', function(assert) {
-  assert.expect(26);
+  assert.expect(27);
 
   this.numbers = numbers;
   this.handleInput = (value, select, e) => {